+++ Editor/Importers/YarnProjectImporter.cs
@@ -12,6 +12,11 @@ using System.IO;
 using UnityEditorInternal;
 using System.Collections;
 
+#if USE_UNITY_LOCALIZATION
+using UnityEditor.Localization;
+using UnityEngine.Localization.Tables;
+#endif
+
 namespace Yarn.Unity.Editor
 {
     [ScriptedImporter(3, new[] { "yarnproject" }, 1), HelpURL("https://yarnspinner.dev/docs/unity/components/yarn-programs/")]
@@ -141,6 +146,11 @@ namespace Yarn.Unity.Editor
         IList<string> IYarnErrorSource.CompileErrors => compileErrors;
 
         bool IYarnErrorSource.Destroyed => this == null;
+
+#if USE_UNITY_LOCALIZATION
+        public bool UseUnityLocalisationSystem = false;
+        public StringTableCollection unityLocalisationStringTableCollection;
+#endif
 
         public override void OnImportAsset(AssetImportContext ctx)
         {
@@ -471,6 +481,13 @@ namespace Yarn.Unity.Editor
                 project.lineMetadata = new LineMetadata(LineMetadataTableEntriesFromCompilationResult(compilationResult));
             }
 
+#if USE_UNITY_LOCALIZATION
+            if (UseUnityLocalisationSystem)
+            {
+                ConvertInternalYarnStringTableEntriesIntoUnityLocalisedStringTableEntries(StringTableEntriesFromCompilationResult(compilationResult));
+            }
+#endif
+
             // Store the compiled program
             byte[] compiledBytes = null;
 
@@ -493,6 +510,40 @@ namespace Yarn.Unity.Editor
 #endif
 
         }
+
+#if USE_UNITY_LOCALIZATION
+        private void ConvertInternalYarnStringTableEntriesIntoUnityLocalisedStringTableEntries(IEnumerable<StringTableEntry> entries)
+        {
+            if (this.unityLocalisationStringTableCollection == null)
+            {
+                Debug.LogError("Unable to generate String Table Entries as the string collection is null");
+                return;
+            }
+
+            var defaultCulture = new System.Globalization.CultureInfo(defaultLanguage);
+
+            foreach (var table in unityLocalisationStringTableCollection.StringTables)
+            {
+                if (table.LocaleIdentifier.CultureInfo != defaultCulture)
+                {
+                    var neutralTable = table.LocaleIdentifier.CultureInfo.IsNeutralCulture ? table.LocaleIdentifier.CultureInfo : table.LocaleIdentifier.CultureInfo.Parent;
+                    var defaultNeutral = defaultCulture.IsNeutralCulture ? defaultCulture : defaultCulture.Parent;
+
+                    if (!neutralTable.Equals(defaultNeutral))
+                    {
+                        continue;
+                    }
+                }
+
+                foreach (var entry in entries)
+                {
+                    table.AddEntry(entry.ID, entry.Text);
+                }
+                return;
+            }
+            Debug.LogWarning($"Unable to find a locale in the string table that matches the default locale {defaultLanguage}");
+        }
+#endif
 
         private List<string> AssemblySearchList()
         {